♻️ Move `TechDebtExtension` to `extension` package for better organization
Relocated the `TechDebtExtension` class to the `extension` package to
align with the module's logical structure.
