+++ techdebt-gradle-plugin/src/main/kotlin/com/escodro/techdebt/gradle/extension/TechDebtExtension.kt
@@ -1,4 +1,4 @@
-package com.escodro.techdebt.gradle
+package com.escodro.techdebt.gradle.extension
 
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
